refactor: separate aircraft and vessel server providers
