+++ src/proxyErrorResponses.test.mjs
@@ -3,7 +3,9 @@ import assert from 'node:assert/strict';
 import { readFileSync } from 'node:fs';
 import path from 'node:path';
 
-const source = readFileSync(new URL('../server/providers/local.js', import.meta.url), 'utf8');
+const source = ['local.js', 'common/http.js', 'aircraft/enrichment.js']
+  .map(file => readFileSync(new URL(`../server/providers/${file}`, import.meta.url), 'utf8'))
+  .join('\n');
 const detail = 'fixture-secret-token /internal/example <html>';
 
 // Execute the production middleware with isolated upstreams and cache storage.
